Check description for changes
Added debug functions for dynamic_management that can be used by a custom allocator.
Created unit test for the API.
Edited tasks.json to add compile options for unit test.
Edited makefile to add compile options for unit test.
number_var comparison (equal) will not be based on integer rounding, instead, uses minimal error on delta (difference between two number).
dynamic_management now overloads new and delete global functions. The overloaded functions will memory management will be based on the configurations.
dynamic_management now includes debug functions and also added macro to help automatically fill in the debug data (only filled if built with DEBUG_BUILD).
New mechanism for allocating and deallocating in dynamic_management.
Added lua::thread for synchronization when using lua_State in multithreading context.
All Lua code that uses lua_lock and lua_unlock will use lua::thread synchronization functions.
Added metadata (stored inside LUA_EXTRASPACE) that can be used directly outside Lua code.
Added functions to get (create a new one if not yet) lua_State based on the current running thread.
Lua API functions will swap the state with thread dependent state internally.
Added functions to temporarily disable thread dependent state mechanism.
Added thread_control and thread_handle for properly handling multithreading with use of Lua API (not the Lua code that is being multithreaded).
runtime_handler run control has been passed to thread_control.
runtime_handler will not store hook_handler and execution_flow as it should exists based on the threads, thread_handle will store them instead.
hook_hanlder is now Thread-Safe.
hook_handler will not be stored as a global variable as of now, it will be stored in a mapped static variable (based on the lua_State address).
execution_flow is now Thread-Safe.
execution_flow will not be stored as a global variable as of now.
variable_watcher is now Thread-Safe (uses CPP Lua global lock).

Fixed bug where memory leaks while executing in runtime_handler (now in thread_handle) thread due to force exiting the thread.

+++ Src/string_util.cpp
@@ -15,7 +15,7 @@ int _strfc(const char* str, size_t strlen, character_filter filter, term_filter 
       break;
 
     if(_tfilter && _tfilter(ch))
-      return SIZE_MAX;
+      return INT_MAX;
   }
 
   return i;
